lib/configuration header: no need to have WITH_GUI and WITH_QWT macros here
The library does not care about the GUI, so there is no point in having
these two macros defined in its header.
WITH_GUI can be removed altogether, since it is an information used by
the build system, to decide if the GUI should be built or not.
WITH_QWT is a GUI property, so set it as a compiler define when Qwt is
found.